[ADF-3336] - unified all the diverging tests into one file
